Match refactored

+++ src/main/java/io/remedymatch/match/api/MatchDTO.java
@@ -1,31 +1,32 @@
 package io.remedymatch.match.api;
 
+import java.math.BigDecimal;
+import java.util.UUID;
+
 import io.remedymatch.institution.api.InstitutionDTO;
 import io.remedymatch.match.domain.MatchStatus;
 import lombok.Builder;
 import lombok.Data;
 
-import java.util.UUID;
-
 @Data
 @Builder
 public class MatchDTO {
 
-    private UUID id;
+	private UUID id;
 
-    private InstitutionDTO institutionVon;
+	private InstitutionDTO institutionVon;
 
-    private MatchStandortDTO standortVon;
+	private MatchStandortDTO standortVon;
 
-    private MatchStandortDTO standortAn;
+	private MatchStandortDTO standortAn;
 
-    private InstitutionDTO institutionAn;
+	private InstitutionDTO institutionAn;
 
-    private UUID anfrageId;
+	private UUID anfrageId;
 
-    private MatchStatus status;
+	private MatchStatus status;
 
-    private String prozessInstanzId;
+	private String prozessInstanzId;
 
-    private double entfernung;
+	private BigDecimal entfernung;
 }
